+++ features/playground-controller.js
@@ -1,18 +1,18 @@
-import { ThemeManager } from '../modules/theme-manager.js?v=20260312-8';
-import { FontManager } from '../modules/font-manager.js?v=20260312-8';
-import { UIManager } from '../modules/ui-manager.js?v=20260312-8';
-import { PanelManager } from '../modules/panel-manager.js?v=20260312-8';
-import { FileManager } from '../modules/file-manager.js?v=20260312-8';
-import { GraphManager } from '../modules/graph-manager.js?v=20260312-8';
-import { PopupManager } from '../modules/popup-manager.js?v=20260312-8';
-import { ClingoManager } from '../modules/clingo-manager.js?v=20260312-8';
-import { OutputManager } from '../modules/output-manager.js?v=20260312-8';
-import { ExportManager } from '../modules/export-manager.js?v=20260312-8';
-import { MetricsManager } from '../modules/metrics-manager.js?v=20260312-8';
-import { ConfigController } from './config-controller.js?v=20260312-8';
-import { DocsController } from './docs-controller.js?v=20260312-8';
-import { EditorController } from './editor-controller.js?v=20260312-8';
-import { ExamplesController } from './examples-controller.js?v=20260312-8';
+import { ThemeManager } from '../modules/theme-manager.js?v=20260312-9';
+import { FontManager } from '../modules/font-manager.js?v=20260312-9';
+import { UIManager } from '../modules/ui-manager.js?v=20260312-9';
+import { PanelManager } from '../modules/panel-manager.js?v=20260312-9';
+import { FileManager } from '../modules/file-manager.js?v=20260312-9';
+import { GraphManager } from '../modules/graph-manager.js?v=20260312-9';
+import { PopupManager } from '../modules/popup-manager.js?v=20260312-9';
+import { ClingoManager } from '../modules/clingo-manager.js?v=20260312-9';
+import { OutputManager } from '../modules/output-manager.js?v=20260312-9';
+import { ExportManager } from '../modules/export-manager.js?v=20260312-9';
+import { MetricsManager } from '../modules/metrics-manager.js?v=20260312-9';
+import { ConfigController } from './config-controller.js?v=20260312-9';
+import { DocsController } from './docs-controller.js?v=20260312-9';
+import { EditorController } from './editor-controller.js?v=20260312-9';
+import { ExamplesController } from './examples-controller.js?v=20260312-9';
 
 export class PlaygroundController {
     constructor(dom, store) {